Remove Templates
We don't need no stinkin' templates.  We're literally just adding some JSON+LD in a block.  The templates are nonsense.  We should obviously just have the parent generate the very small amount of HTML thats needed and extend that with the information we need to fill in.

This results in far cleaner code and less necessary files.